requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# LangGraph and LangChain
langchain
langgraph
langchain-community
langchain-google-genai




# Embedding and Vector Database
sentence-transformers==2.2.2
elasticsearch==8.11.0
numpy==1.24.3

# HTTP and API
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
python-dotenv==1.0.0

# UI Components
plotly==5.17.0
streamlit-option-menu==0.3.6

# Google AI
google-generativeai==0.3.2